Implemented Logic for Bridge
TODO ACK information handling
TODO Task logic
TODO If OS is write --> if return is null don't add etc

+++ io.openems.edge.bridge.genibus/src/io/openems/edge/bridge/genibus/protocol/ApplicationProgramDataUnit.java
@@ -11,7 +11,7 @@ public class ApplicationProgramDataUnit {
     // 0...63 Byte
     ByteArrayOutputStream apduDataFields = new ByteArrayOutputStream();
 
-    public short getHead() {
+    public int getHead() {
         return apduHead;
     }
